// sanity/lib/interfaces.ts
export interface PostCard {
    title: string;
    slug: string;
    publishedAt: string;
    category?: string;
    excerpt?: string;
    imageUrl?: string;
    author?: string;
    words?: number;
};

export interface Post extends PostCard {
    _updatedAt: string;
    body?: PortableTextBlock[];
    authorPosition?: string;
    authorImageUrl?: string;
};